+++ tests/snappi_tests/bgp/test_bgp_outbound_uplink_po_member_flap.py
@@ -5,8 +5,8 @@ from tests.common.fixtures.conn_graph_facts import conn_graph_facts, \
      fanout_graph_facts_multidut                                                                     # noqa: F401
 from tests.common.snappi_tests.snappi_fixtures import snappi_api_serv_ip, snappi_api_serv_port, \
      snappi_api, multidut_snappi_ports_for_bgp                                                       # noqa: F401
-from tests.snappi_tests.variables import t1_t2_device_hostnames                        # noqa: F401
-from tests.snappi_tests.multidut.bgp.files.bgp_outbound_helper import (
+from tests.snappi_tests.variables import t1_t2_device_hostnames                         # noqa: F401
+from tests.snappi_tests.bgp.files.bgp_outbound_helper import (
      run_bgp_outbound_link_flap_test)                                                               # noqa: F401
 from tests.common.snappi_tests.snappi_test_params import SnappiTestParams                           # noqa: F401
 
@@ -42,14 +42,14 @@ ROUTE_RANGES = [{
             }]
 
 
-def test_bgp_outbound_uplink_po_flap(snappi_api,                                     # noqa: F811
-                                     multidut_snappi_ports_for_bgp,                       # noqa: F811
-                                     conn_graph_facts,                             # noqa: F811
-                                     fanout_graph_facts_multidut,                   # noqa: F811
-                                     duthosts,
-                                     creds):
+def test_bgp_outbound_uplink_po_member_flap(snappi_api,                                     # noqa: F811
+                                            multidut_snappi_ports_for_bgp,                # noqa: F811
+                                            conn_graph_facts,                             # noqa: F811
+                                            fanout_graph_facts_multidut,                   # noqa: F811
+                                            duthosts,
+                                            creds):
     """
-    Gets the packet loss duration on flapping portchannel in uplink side
+    Gets the packet loss duration on flapping portchannel member in uplink side
 
     Args:
         snappi_api (pytest fixture): SNAPPI session
@@ -64,7 +64,7 @@ def test_bgp_outbound_uplink_po_flap(snappi_api,                                
     snappi_extra_params = SnappiTestParams()
     snappi_extra_params.ROUTE_RANGES = ROUTE_RANGES
     snappi_extra_params.iteration = ITERATION
-    snappi_extra_params.test_name = "T2 Uplink Portchannel Flap"
+    snappi_extra_params.test_name = "T2 Uplink Portchannel Member Flap"
     snappi_extra_params.multi_dut_params.flap_details = FLAP_DETAILS
 
     if (len(t1_t2_device_hostnames) < 3) or (len(duthosts) < 3):
@@ -88,7 +88,6 @@ def test_bgp_outbound_uplink_po_flap(snappi_api,                                
             snappi_extra_params.multi_dut_params.duthost3 = duthost
         else:
             continue
-
     snappi_extra_params.multi_dut_params.multi_dut_ports = multidut_snappi_ports_for_bgp
     run_bgp_outbound_link_flap_test(api=snappi_api,
                                     creds=creds,
