+++ tests/snappi_tests/bgp/test_bgp_outbound_uplink_process_crash.py
@@ -2,11 +2,11 @@ import pytest
 import logging
 from tests.common.helpers.assertions import pytest_require, pytest_assert                            # noqa: F401
 from tests.common.fixtures.conn_graph_facts import conn_graph_facts, \
-     fanout_graph_facts_multidut                                                                    # noqa: F401
+     fanout_graph_facts_multidut                                                                     # noqa: F401
 from tests.common.snappi_tests.snappi_fixtures import snappi_api_serv_ip, snappi_api_serv_port, \
-     snappi_api, multidut_snappi_ports_for_bgp                                                      # noqa: F401
+     snappi_api, multidut_snappi_ports_for_bgp                                                       # noqa: F401
 from tests.snappi_tests.variables import t1_t2_device_hostnames                                     # noqa: F401
-from tests.snappi_tests.multidut.bgp.files.bgp_outbound_helper import (
+from tests.snappi_tests.bgp.files.bgp_outbound_helper import (
      run_bgp_outbound_process_restart_test)                                                          # noqa: F401
 from tests.common.snappi_tests.snappi_test_params import SnappiTestParams                           # noqa: F401
 
@@ -37,14 +37,14 @@ ROUTE_RANGES = [{
             }]
 
 
-def test_bgp_outbound_downlink_process_crash(snappi_api,                                     # noqa: F811
-                                             multidut_snappi_ports_for_bgp,                       # noqa: F811
-                                             conn_graph_facts,                             # noqa: F811
-                                             fanout_graph_facts_multidut,                           # noqa: F811
-                                             duthosts,                                  # noqa: F811
-                                             creds):                                # noqa: F811
+def test_bgp_outbound_uplink_process_crash(snappi_api,                                     # noqa: F811
+                                           multidut_snappi_ports_for_bgp,                  # noqa: F811
+                                           conn_graph_facts,                             # noqa: F811
+                                           fanout_graph_facts_multidut,                  # noqa: F811
+                                           duthosts,                                # noqa: F811
+                                           creds):                                # noqa: F811
     """
-    Gets the packet loss duration on killing certain processes in downlink
+    Gets the packet loss duration on killing certain processes in uplink
 
     Args:
         snappi_api (pytest fixture): SNAPPI session
@@ -59,12 +59,12 @@ def test_bgp_outbound_downlink_process_crash(snappi_api,                        
     snappi_extra_params = SnappiTestParams()
     snappi_extra_params.ROUTE_RANGES = ROUTE_RANGES
     snappi_extra_params.iteration = ITERATION
-    snappi_extra_params.test_name = "T2 Downlink Process Crash"
+    snappi_extra_params.test_name = "T2 Uplink Process Crash"
     snappi_extra_params.multi_dut_params.process_names = {
                                                             'swss': "/usr/bin/orchagent",
                                                             'syncd': "/usr/bin/syncd",
                                                         }
-    snappi_extra_params.multi_dut_params.host_name = t1_t2_device_hostnames[2]
+    snappi_extra_params.multi_dut_params.host_name = t1_t2_device_hostnames[1]
     if (len(t1_t2_device_hostnames) < 3) or (len(duthosts) < 3):
         pytest_require(False, "Need minimum of 3 devices : One T1 and Two T2 line cards")
